% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Maia.r
\name{opiClose_for_Maia}
\alias{opiClose_for_Maia}
\title{Implementation of opiClose for the Maia machine.}
\arguments{
\item{\code{}}{A list containing:}
}
\value{
A list containing:
\itemize{
\item \code{res} List of result elements.
\item \code{res.msg} The error message or additional results from the CLOSE command
\item \code{res.error} '0' if success, something else if error.
\item \code{res.time} The time stamp for fixation data
\item \code{res.x} The time stamp for fixation data
\item \code{res.y} The time stamp for fixation data
}
}
\description{
This is for internal use only. Use \code{\link[=opiClose]{opiClose()}} with
these Arguments and you will get the Value back.
}
\examples{
chooseOpi("Maia")
result <- opiClose()

}
\seealso{
\code{\link[=opiClose]{opiClose()}}
}
