% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/O900.r
\name{opiClose_for_O900}
\alias{opiClose_for_O900}
\title{Implementation of opiClose for the O900 machine.}
\arguments{
\item{\code{}}{A list containing:}
}
\value{
A list containing:
\itemize{
\item \code{res} List of result elements.
\item \code{res.msg} The error message or additional results from the CLOSE command
\item \code{res.error} '0' if success, something else if error.
}
}
\description{
This is for internal use only. Use \code{\link[=opiClose]{opiClose()}} with
these Arguments and you will get the Value back.
}
\examples{
chooseOpi("O900")
result <- opiClose()

}
\seealso{
\code{\link[=opiClose]{opiClose()}}
}
