% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimYes.r
\name{opiClose_for_SimYes}
\alias{opiClose_for_SimYes}
\title{Does nothing.}
\usage{
opiClose_for_SimYes()
}
\value{
A list with elements:
\itemize{
\item \code{error} Always \code{FALSE}.
\item \code{msg} A string "Close OK".
}
}
\description{
Does nothing.
}
