% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Maia.r
\name{opiInitialise_for_Maia}
\alias{opiInitialise_for_Maia}
\title{Implementation of opiInitialise for the Maia machine.}
\arguments{
\item{\code{address}}{A list containing:
\itemize{
\item \code{port} TCP port of the OPI Monitor.
\item \code{ip} IP Address of the OPI Monitor.
}}
}
\value{
A list containing:
\itemize{
\item \code{error} \code{TRUE} if there was an error, \code{FALSE} if not.
\item \code{msg} If \code{error} is \code{TRUE}, then this is a string describing the error.
If \code{error} is \code{FALSE}, this is a list of:
\itemize{
\item \code{image} The captured IR image at initialisation (base64 encoded).
\item \code{onhy} The y-coordinate of the ONH measured initialisation (pixels).
\item \code{prlx} The x-coordinate of the PRL measured initialisation (pixels).
\item \code{prly} The y-coordinate of the PRL measured initialisation (pixels).
\item \code{onhx} The x-coordinate of the ONH measured initialisation (pixels).
}
}
}
\description{
This is for internal use only. Use \code{\link[=opiInitialise]{opiInitialise()}} with
these Arguments and you will get the Value back.
}
\details{
\code{port} can take on values in the range \code{[0, 65535]}.
}
\examples{
chooseOpi("Maia")
result <- opiInitialise(address = list(port = 50001, ip = "localhost"))

}
\seealso{
\code{\link[=opiInitialise]{opiInitialise()}}
}
