% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/O900.r
\name{opiInitialise_for_O900}
\alias{opiInitialise_for_O900}
\title{Implementation of opiInitialise for the O900 machine.}
\arguments{
\item{\code{address}}{A list containing:
\itemize{
\item \code{port} TCP port of the OPI Monitor.
\item \code{ip} IP Address of the OPI Monitor.
}}
}
\value{
A list containing:
\itemize{
\item \code{res} List with all of the other fields described in @ReturnMsg except 'error'.
\item \code{res.msg} The success or error message.
\item \code{res.error} Error code '0' if all good, something else otherwise.
}
}
\description{
This is for internal use only. Use \code{\link[=opiInitialise]{opiInitialise()}} with
these Arguments and you will get the Value back.
}
\details{
\code{port} can take on values in the range \code{[0, 65535]}.
}
\examples{
chooseOpi("O900")
result <- opiInitialise(address = list(port = 50001, ip = "localhost"))

}
\seealso{
\code{\link[=opiInitialise]{opiInitialise()}}
}
