% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimGaussian.r
\name{opiInitialise_for_SimGaussian}
\alias{opiInitialise_for_SimGaussian}
\title{Simulates responses using a Frequency of Seeing (FoS) curve.}
\usage{
opiInitialise_for_SimGaussian(sd = 1, maxStim = 10000/pi, ...)
}
\arguments{
\item{sd}{Standard deviation of Cumulative Gaussian.}

\item{maxStim}{The maximum stimuls value (0 dB) in cd/\eqn{\mbox{m}^2}{m^2}.}

\item{...}{Any other parameters you like, they are ignored.}
}
\value{
A list with elements:
\itemize{
\item \code{error} \code{FALSE} if machine initialised, \code{TRUE} otherwise.
\item \code{msg} A string message.
}
}
\description{
The FoS is modelled as a cumulative Gaussian function with standard deviation
equal to \code{sd} as provided and the mean as the true threshold given as \code{tt}
\link{opiPresent}.
All values are in dB relative to \code{maxStim}.
}
\examples{
    # Set up a simple simulation for white-on-white perimetry
chooseOpi("SimGaussian")
res <- opiInitialize(sd = 2.5)
if (res$error)
  stop(paste("opiInitialize() failed:", res$msg))

}
