% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimHenson.r
\name{opiInitialise_for_SimHenson}
\alias{opiInitialise_for_SimHenson}
\title{Simulates reponses using a Frequency of Seeing (FoS) curve.}
\usage{
opiInitialise_for_SimHenson(
  type = "C",
  A = -0.081,
  B = 3.27,
  cap = 6,
  maxStim = 10000/pi,
  ...
)
}
\arguments{
\item{type}{A single character that is:
\itemize{
\item \code{N} for using the A and B values from the Normals in Henson et al (2000)
\item \code{G} for using the A and B values from the Glaucomas in Henson et al (2000)
\item \code{C} for using the A and B values from the Combined in Henson et al (2000)
\item \code{X} to specify your own A and B values as parameters
}}

\item{A}{Coefficient of \code{t} in the formula (ignored if \code{type != 'X'}).}

\item{B}{Addend of \code{t} in the formula (ignored if \code{type != 'X'}).}

\item{cap}{Maximum dB value for the stdev of the FoS curve.}

\item{maxStim}{The maximum stimuls value (0 dB) in cd/\eqn{\mbox{m}^2}{m^2}.}

\item{...}{Any other parameters you like, they are ignored.}
}
\value{
NULL on success or a string message otherwise
}
\description{
The FoS is modelled as a cummulative Gaussian function over dB with
standard deviation equal to \code{min(cap, exp( A * t + B))}, where
t is the threshold/mean of the FoS in dB.
All values are in dB relative to \code{maxStim}.
}
\examples{
    # Set up a simple simulation for white-on-white perimetry
chooseOpi("SimHenson")
if (!is.null(opiInitialize(type = "C", cap = 6)))
  stop("opiInitialize failed")

}
