% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimYes.r
\name{opiInitialise_for_SimYes}
\alias{opiInitialise_for_SimYes}
\title{Does nothing.}
\usage{
opiInitialise_for_SimYes(...)
}
\arguments{
\item{...}{Any other paramters you like, they are ignored.}
}
\description{
Does nothing.
}
