% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PicoVR.r
\name{opiPresent_for_PicoVR}
\alias{opiPresent_for_PicoVR}
\title{Implementation of opiPresent for the PicoVR machine.}
\arguments{
\item{\code{stim}}{A list containing:
\itemize{
\item \code{phase} List of phases (in degrees) for generation of spatial
patterns. Only useful if type != FLAT (Optional)
\item \code{imageFilename} If type == IMAGE, the filename on the local
filesystem of the machine running JOVP of the image to use (Optional)
\item \code{shape} Stimulus shape. Values include CROSS, TRIANGLE, CIRCLE,
SQUARE, OPTOTYPE. (Optional)
\item \code{sx} List of diameters along major axis of ellipse (degrees).
\item \code{lum} List of stimuli luminances (cd/m^2).
\item \code{sy} List of diameters along minor axis of ellipse (degrees). If not
received, then sy = sx (Optional)
\item \code{rotation} List of angles of rotation of stimuli (degrees). Only
useful if sx != sy specified. (Optional)
\item \code{texRotation} List of angles of rotation of stimuli (degrees). Only
useful if type != FLAT (Optional)
\item \code{type} Stimulus type. Values include FLAT, SINE, CHECKERBOARD,
SQUARESINE, G1, G2, G3, IMAGE (Optional)
\item \code{stim.length} The number of elements in this stimuli.
\item \code{defocus} List of defocus values in Diopters for stimulus post-processing. (Optional)
\item \code{frequency} List of frequencies (in cycles per degrees) for
generation of spatial patterns. Only useful if type != FLAT (Optional)
\item \code{eye} The eye for which to apply the settings.
\item \code{color1} List of stimulus colors for FLAT shapes and patterns.
\item \code{color2} List of second colors for non-FLAT shapes (Optional)
\item \code{t} List of stimuli presentation times (ms).
\item \code{w} Time to wait for response including presentation time (ms).
\item \code{contrast} List of stimulus contrasts (from 0 to 1). Only useful if
type != FLAT. (Optional)
\item \code{optotype} If shape == OPTOTYPE, the letter A to Z to use (Optional)
\item \code{x} List of x co-ordinates of stimuli (degrees).
\item \code{y} List of y co-ordinates of stimuli (degrees).
}}

\item{\code{...}}{Parameters for other opiPresent implementations that are ignored here.}
}
\value{
A list containing:
\itemize{
\item \code{res} List with all of the other fields described in @ReturnMsg except 'error'.
\item \code{res.msg.eyey} y co-ordinates of pupil at times eyet (degrees).
\item \code{res.msg.eyex} x co-ordinates of pupil at times eyet (degrees).
\item \code{res.msg.time} Response time from stimulus onset if button pressed (ms).
\item \code{res.msg.eyed} Diameter of pupil at times eyet (mm).
\item \code{res.msg.eyet} Time of (eyex, eyey) pupil from stimulus onset (ms).
\item \code{res.msg} Error message or a structure with the following fields.
\item \code{res.error} '0' if success, something else if error.
\item \code{res.msg.seen} '1' if seen, '0' if not.
}
}
\description{
This is for internal use only. Use \code{\link[=opiPresent]{opiPresent()}} with
these Arguments and you will get the Value back.
}
\details{
Elements in \code{phase} can take on values in the range
\code{[0.0, 1.0E10]}.

Elements in \code{shape} can take on values in the set
\code{{"triangle", "square", "polygon", "hollow_triangle",
                  "hollow_square", "hollow_polygon", "cross", "maltese",
                  "circle", "annulus", "optotype", "text", "model"}}.

Elements in \code{sx} can take on values in the range \code{[0.0, 180.0]}.

Elements in \code{lum} can take on values in the range \code{[0.0, 1.0E10]}.

Elements in \code{sy} can take on values in the range \code{[0.0, 180.0]}.

Elements in \code{rotation} can take on values in the
range \code{[0.0, 360.0]}.

Elements in \code{texRotation} can take on values
in the range \code{[0.0, 360.0]}.

Elements in \code{type} can take on values in the set
\code{{"flat", "checkerboard", "sine", "squaresine", "g1",
                 "g2", "g3", "text", "image"}}.

\code{stim.length} can take on values in the range \code{[1, 2147483647]}.

Elements in \code{defocus} can take on values in the
range \code{[0.0, 1.0E10]}.

Elements in \code{frequency} can take on values in
the range \code{[0.0, 300.0]}.

Elements in \code{eye} can take on values in the set
\code{{"left", "right", "both", "none"}}.

Elements in \code{color1} can take on values in the range \code{[0.0, 1.0]}.

Elements in \code{color2} can take on values in the range \code{[0.0, 1.0]}.

Elements in \code{t} can take on values in the range \code{[0.0, 1.0E10]}.

\code{w} can take on values in the range \code{[0.0, 1.0E10]}.

Elements in \code{contrast} can take on values in the
range \code{[0.0, 1.0]}.

Elements in \code{optotype} can take on values in the
set \code{{"a", "b", "c", "d", "e", "f", "g", "h", "i",
                     "j", "k", "l", "m", "n", "o", "p", "q", "r", "s", "t",
                     "u", "v", "w", "x", "y", "z"}}.

Elements in \code{x} can take on values in the range \code{[-90.0, 90.0]}.

Elements in \code{y} can take on values in the range \code{[-90.0, 90.0]}.
}
\examples{
chooseOpi("PicoVR")
result <- opiPresent(stim = list(sx = list(1.72), lum = list(300.0), stim.length = 1,
                  eye = list("LEFT"), color1 = list(list(0.0, 0.0, 0.0)),
                  t = list(200.0), w = 1500.0, x = list(0.0), y = list(0.0)))

}
\seealso{
\code{\link[=opiPresent]{opiPresent()}}
}
