% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimHenson.r
\name{opiPresent_for_SimHenson}
\alias{opiPresent_for_SimHenson}
\title{Determine the response to a stimuli by sampling from a cumulative Gaussian
Frequency-of-Seeing (FoS) curve (also known as the psychometric function).}
\usage{
opiPresent_for_SimHenson(stim, fpr = 0.03, fnr = 0.01, tt = 30, ...)
}
\arguments{
\item{stim}{A list that contains at least:
\itemize{
\item \code{lum} which is the stim value in cd/\eqn{\mbox{m}^2}{m^2}.
}}

\item{fpr}{false positive rate for the FoS curve (range 0..1).}

\item{fnr}{false negative rate for the FoS curve (range 0..1).}

\item{tt}{mean of the assumed FoS curve in dB.}

\item{...}{Any other parameters you like, they are ignored.}
}
\value{
A list with elements:
\itemize{
\item \code{error} \code{TRUE} if error, \code{FALSE} otherwise.
\item \code{msg} A string if \code{error} is \code{TRUE} else a list containing
\itemize{
\item \code{seen} \code{TRUE} or \code{FALSE}.
\item \code{time} Always \code{NA}.
}
}
}
\description{
The FoS formula is
\deqn{\mbox{fpr}+(1-\mbox{fpr}-\mbox{fnr})(1-\mbox{pnorm}(x, \mbox{tt}, \mbox{pxVar})}
where \code{x} is the stimulus value in dB, and \code{pxVar} is
\deqn{\min(\mbox{cap}, e^{A\times\mbox{tt}+B}).}
The ceiling \code{cap} is set with the call to
\code{opiInitialize}, and \code{A} and \code{B} are from Table 1 in Henson
et al (2000), also set in the call to \code{opiInitiaise} using the \code{type} parameter.
}
\examples{
    # Stimulus is Size III white-on-white as in the HFA
chooseOpi("SimHenson")
res <- opiInitialize(type = "C", cap = 6)
if (res$error)
  stop(paste("opiInitialize() failed:", res$msg))

result <- opiPresent(stim = list(lum = dbTocd(20)), tt = 30, fpr = 0.15, fnr = 0.01)

res <- opiClose()
if (res$error)
  warning(paste("opiClose() failed:", res$msg))

}
