% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimNo.r
\name{opiPresent_for_SimNo}
\alias{opiPresent_for_SimNo}
\title{Always respond 'not seen' to any parameter.
No checking is done on the validity of \code{stim}.}
\usage{
opiPresent_for_SimNo(...)
}
\arguments{
\item{...}{Any parameters you like, they are ignored.}
}
\value{
A list with elements:
\itemize{
\item \code{error} Always \code{FALSE}.
\item \code{msg} A list containing
\itemize{
\item \code{seen} Always \code{FALSE}.
\item \code{time} Always \code{NA}.
}
}
}
\description{
Always respond 'not seen' to any parameter.
No checking is done on the validity of \code{stim}.
}
