% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/O900.r
\name{opiQueryDevice_for_O900}
\alias{opiQueryDevice_for_O900}
\title{Implementation of opiQueryDevice for the O900 machine.}
\arguments{
\item{\code{}}{A list containing:}
}
\value{
A list containing:
\itemize{
\item \code{res} List with all of the other fields described in @ReturnMsg except 'error'.
\item \code{res.msg} The error message or a structure with the following data.
\item \code{res.error} '0' if success, something else if error.
}
}
\description{
This is for internal use only. Use \code{\link[=opiQueryDevice]{opiQueryDevice()}} with
these Arguments and you will get the Value back.
}
\examples{
chooseOpi("O900")
result <- opiQueryDevice()

}
\seealso{
\code{\link[=opiQueryDevice]{opiQueryDevice()}}
}
