% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimGaussian.r
\name{opiQueryDevice_for_SimGaussian}
\alias{opiQueryDevice_for_SimGaussian}
\title{Returns a simple list.}
\usage{
opiQueryDevice_for_SimGaussian()
}
\value{
A list with one element \code{machine} that is \code{"SimHenson"}.
}
\description{
Returns a simple list.
}
