% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimHenson.r
\name{opiQueryDevice_for_SimHenson}
\alias{opiQueryDevice_for_SimHenson}
\title{Returns name of the machine.}
\usage{
opiQueryDevice_for_SimHenson()
}
\value{
A list with elements:
\itemize{
\item \code{error} Always \code{FALSE}.
\item \code{msg} A list containing \code{machine} that is set to \code{"SimHenson"}.
}
}
\description{
Returns name of the machine.
}
