% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimNo.r
\name{opiQueryDevice_for_SimNo}
\alias{opiQueryDevice_for_SimNo}
\title{Returns name of the machine.}
\usage{
opiQueryDevice_for_SimNo()
}
\value{
A list with elements:
\itemize{
\item \code{error} Always \code{FALSE}.
\item \code{msg} A list containing \code{machine} that is set to \code{"SimNo"}.
}
}
\description{
Returns name of the machine.
}
