% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimYes.r
\name{opiQueryDevice_for_SimYes}
\alias{opiQueryDevice_for_SimYes}
\title{Returns a simple list.}
\usage{
opiQueryDevice_for_SimYes()
}
\value{
A list with one element \code{machine} that is \code{"SimYes"}.
}
\description{
Returns a simple list.
}
