% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Display.r
\name{opiSetup_for_Display}
\alias{opiSetup_for_Display}
\title{Implementation of opiSetup for the Display machine.}
\arguments{
\item{\code{settings}}{A list containing:
\itemize{
\item \code{eye} The eye for which to apply the settings.#'  * \code{bgImageFilename} If present, display the image in the background for eye (Optional)
\item \code{fixShape} Fixation target type for eye. (Optional)
\item \code{fixLum} Fixation target luminance for eye. (Optional)
\item \code{fixType} Fixation target texture for eye. (Optional)
\item \code{fixCx} x-coordinate of fixation target (degrees). (Optional)
\item \code{fixCy} y-coordinate of fixation target (degrees). (Optional)
\item \code{fixCol} Fixation target color for eye. (Optional)
\item \code{bgLum} Background luminance for eye (cd/m^2). (Optional)
\item \code{tracking} Whether to correct stimulus location based on eye position. (Optional)
\item \code{bgCol} Background color for eye (rgb). (Optional)
\item \code{fixSx} diameter along major axis of ellipse (degrees). 0 to hide
fixation marker. (Optional)
\item \code{fixSy} diameter along minor axis of ellipse (degrees). If not
received, then sy = sx. (Optional)
\item \code{fixRotation} Angles of rotation of fixation target (degrees). Only
useful if sx != sy specified. (Optional)
\item \code{fixImageFilename} If fixType == IMAGE, the filename on the local
filesystem of the machine running JOVP of the image to use (Optional)
}}
}
\value{
A list containing:
\itemize{
\item \code{error} \code{TRUE} if there was an error, \code{FALSE} if not.
\item \code{msg} If \code{error} is \code{TRUE}, then this is a string describing the error.
If \code{error} is \code{FALSE}, this is an empty list.
}
}
\description{
This is for internal use only. Use \code{\link[=opiSetup]{opiSetup()}} with
these Arguments and you will get the Value back.
}
\details{
\code{fixShape} can take on values in the set \code{{"triangle",
          "square", "polygon", "hollow_triangle", "hollow_square",
          "hollow_polygon", "cross", "maltese", "circle", "annulus",
          "optotype", "text", "model"}}.

\code{fixLum} can take on values in the range \code{[0.0, 1.0E10]}.

\code{fixType} can take on values in the set \code{{"flat",
         "checkerboard", "sine", "squaresine", "g1", "g2", "g3", "text", "image"}}.

\code{fixCx} can take on values in the range \code{[-90.0, 90.0]}.

\code{fixCy} can take on values in the range \code{[-90.0, 90.0]}.

Elements in \code{fixCol} can take on values in the range \code{[0.0, 1.0]}.

\code{bgLum} can take on values in the range \code{[0.0, 1.0E10]}.

\code{tracking} can take on values in the range \code{[0, 1]}.

Elements in \code{bgCol} can take on values in the range \code{[0.0, 1.0]}.

\code{eye} can take on values in the set \code{{"left", "right",
     "both", "none"}}.

\code{fixSx} can take on values in the range \code{[0.0, 1.0E10]}.

\code{fixSy} can take on values in the range \code{[0.0, 1.0E10]}.

\code{fixRotation} can take on values in the range \code{[0.0, 360.0]}.
}
\examples{
chooseOpi("Display")
result <- opiSetup(settings = list(eye = "BOTH"))

}
\seealso{
\code{\link[=opiSetup]{opiSetup()}}
}
