% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/O900.r
\name{opiSetup_for_O900}
\alias{opiSetup_for_O900}
\title{Implementation of opiSetup for the O900 machine.}
\arguments{
\item{\code{settings}}{A list containing:
\itemize{
\item \code{eye} Eye to set.
\item \code{fixShape} Fixation target.
\item \code{pres} Volume for auditory feedback when a stimulus is presented: 0 means no buzzer.
\item \code{resp} Volume for auditory feedback when observer presses the
clicker: 0 means no buzzer.
\item \code{fixIntensity} Fixation intensity(from 0\% to 100\%).
\item \code{bgLum} Background luminance for eye.
\item \code{bgCol} Background color for eye.
}}
}
\value{
A list containing:
\itemize{
\item \code{res} List with all of the other fields described in @ReturnMsg except 'error'.
\item \code{res.msg} The error message or a structure with the result of QUERY OPI command.
\item \code{res.error} '0' if success, something else if error.
}
}
\description{
This is for internal use only. Use \code{\link[=opiSetup]{opiSetup()}} with
these Arguments and you will get the Value back.
}
\details{
\code{eye} can take on values in the set \code{{"left", "right"}}.

\code{fixShape} can take on values in the set \code{{"center",
          "cross", "ring"}}.

\code{pres} can take on values in the range \code{[0.0, 3.0]}.

\code{resp} can take on values in the range \code{[0.0, 3.0]}.

\code{fixIntensity} can take on values in the range \code{[0.0, 100.0]}.

\code{bgLum} can take on values in the set \code{{"bg_off", "bg_1",
       "bg_10", "bg_100"}}.

\code{bgCol} can take on values in the set \code{{"white", "yellow"}}.
}
\examples{
chooseOpi("O900")
result <- opiSetup(settings = list(eye = "null", fixShape = "null", pres = 0.0, resp = 0.0,
                fixIntensity = 50.0, bgLum = "null", bgCol = "null"))

}
\seealso{
\code{\link[=opiSetup]{opiSetup()}}
}
