% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PicoVR.r
\name{opiSetup_for_PicoVR}
\alias{opiSetup_for_PicoVR}
\title{Implementation of opiSetup for the PicoVR machine.}
\arguments{
\item{\code{settings}}{A list containing:
\itemize{
\item \code{bgImageFilename} If present, display the image in the background for eye (Optional)
\item \code{fixShape} Fixation target type for eye.
\item \code{fixLum} Fixation target luminance for eye.
\item \code{fixCx} x-coordinate of fixation target (degrees).
\item \code{fixCy} y-coordinate of fixation target (degrees).
\item \code{fixCol} Fixation target color for eye.
\item \code{bgLum} Background luminance for eye (cd/m^2).
\item \code{tracking} Whether to correct stimulus location based on eye position. (Optional)
\item \code{bgCol} Background color for eye (rgb).
\item \code{eye} The eye for which to apply the settings.
\item \code{fixSx} diameter along major axis of ellipse (degrees).
\item \code{fixSy} diameter along minor axis of ellipse (degrees). If not
received, then sy = sx. (Optional)
\item \code{fixRotation} Angles of rotation of fixation target (degrees). Only
useful if sx != sy specified. (Optional)
}}
}
\value{
A list containing:
\itemize{
\item \code{res} List with all of the other fields described in @ReturnMsg except 'error'.
\item \code{res.msg} The error message or a structure with the result of QUERY OPI command.
\item \code{res.error} '0' if success, something else if error.
}
}
\description{
This is for internal use only. Use \code{\link[=opiSetup]{opiSetup()}} with
these Arguments and you will get the Value back.
}
\details{
\code{fixShape} can take on values in the set \code{{"triangle",
          "square", "polygon", "hollow_triangle", "hollow_square",
          "hollow_polygon", "cross", "maltese", "circle", "annulus",
          "optotype", "text", "model"}}.

\code{fixLum} can take on values in the range \code{[0.0, 1.0E10]}.

\code{fixCx} can take on values in the range \code{[-90.0, 90.0]}.

\code{fixCy} can take on values in the range \code{[-90.0, 90.0]}.

Elements in \code{fixCol} can take on values in the range \code{[0.0, 1.0]}.

\code{bgLum} can take on values in the range \code{[0.0, 1.0E10]}.

\code{tracking} can take on values in the range \code{[0, 1]}.

Elements in \code{bgCol} can take on values in the range \code{[0.0, 1.0]}.

\code{eye} can take on values in the set \code{{"left", "right",
     "both", "none"}}.

\code{fixSx} can take on values in the range \code{[0.0, 1.0E10]}.

\code{fixSy} can take on values in the range \code{[0.0, 1.0E10]}.

\code{fixRotation} can take on values in the range \code{[0.0, 360.0]}.
}
\examples{
chooseOpi("PicoVR")
result <- opiSetup(settings = list(fixShape = "MALTESE", fixLum = 20.0, fixCx = 0.0, fixCy = 0.0,
                fixCol = list(0.0, 1.0, 0.0), bgLum = 10.0, bgCol = list(1.0,
                1.0, 1.0), eye = "BOTH", fixSx = 1.0))

}
\seealso{
\code{\link[=opiSetup]{opiSetup()}}
}
