% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Maia.r
\name{opiClose_for_Maia}
\alias{opiClose_for_Maia}
\title{Implementation of opiClose for the Maia machine.}
\arguments{
\item{\code{}}{A list containing:}
}
\value{
A list containing:
\itemize{
\item \code{error} \code{TRUE} if there was an error, \code{FALSE} if not.
\item \code{msg} If \code{error} is \code{TRUE}, then this is a string describing the error.
If \code{error} is \code{FALSE}, this is a list of:
\itemize{
\item \code{x} The time stamp for fixation data
\item \code{y} The time stamp for fixation data
\item \code{time} The time stamp for fixation data
}
}
}
\description{
This is for internal use only. Use \code{\link[=opiClose]{opiClose()}} with
these Arguments and you will get the Value back.
}
\examples{
chooseOpi("Maia")
result <- opiClose()

}
\seealso{
\code{\link[=opiClose]{opiClose()}}
}
