% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Maia.r
\name{opiInitialise_for_Maia}
\alias{opiInitialise_for_Maia}
\title{Implementation of opiInitialise for the Maia machine.}
\arguments{
\item{\code{address}}{A list containing:
\itemize{
\item \code{port} TCP port of the OPI Monitor.
\item \code{ip} IP Address of the OPI Monitor.
}}
}
\value{
A list containing:
\itemize{
\item \code{res} List with all of the other fields described in @ReturnMsg except 'error'.
\item \code{res.prly} The y-coordinate of the PRL measured initialisation (pixels).
\item \code{res.onhx} The x-coordinate of the ONH measured initialisation (pixels).
\item \code{res.prlx} The x-coordinate of the PRL measured initialisation (pixels).
\item \code{res.msg} The success or error message.
\item \code{res.error} Error code '0' if all good, something else otherwise.
\item \code{res.onhy} The y-coordinate of the ONH measured initialisation (pixels).
\item \code{res.image} The captured IR image at initialisation (base64 encoded).
}
}
\description{
This is for internal use only. Use \code{\link[=opiInitialise]{opiInitialise()}} with
these Arguments and you will get the Value back.
}
\details{
\code{port} can take on values in the range \code{[0, 65535]}.
}
\examples{
chooseOpi("Maia")
result <- opiInitialise(address = list(port = 50001, ip = "localhost"))

}
\seealso{
\code{\link[=opiInitialise]{opiInitialise()}}
}
