% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimGaussian.r
\name{opiInitialise_for_SimGaussian}
\alias{opiInitialise_for_SimGaussian}
\title{Simulates reponses using a Frequency of Seeing (FoS) curve.}
\usage{
opiInitialise_for_SimGaussian(sd = 1, maxStim = 10000/pi, ...)
}
\arguments{
\item{sd}{Standard deviation of Cummulative Gaussian.}

\item{maxStim}{The maximum stimuls value (0 dB) in cd/\eqn{\mbox{m}^2}{m^2}.}

\item{...}{Any other parameters you like, they are ignored.}
}
\value{
NULL on success or a string message otherwise
}
\description{
The FoS is modelled as a cummulative Gaussian function with standard deviation
equal to \code{sd} as provided and the mean as the true threshold given as \code{tt}
\link{opiPresent}.
All values are in dB relative to \code{maxStim}.
}
\examples{
    # Set up a simple simulation for white-on-white perimetry
chooseOpi("SimGaussian")
if (!is.null(opiInitialize(sd = 2.5)))
  stop("opiInitialize failed")

}
