% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimYes.r
\name{opiInitialise_for_SimYes}
\alias{opiInitialise_for_SimYes}
\title{Does nothing.}
\usage{
opiInitialise_for_SimYes(...)
}
\arguments{
\item{...}{Any object you like, it is ignored.}
}
\value{
A list with elements:
\itemize{
\item \code{error} Always \code{FALSE}.
\item \code{msg} A string "Initialise OK"
}
}
\description{
Does nothing.
}
