% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compass.r
\name{opiPresent_for_Compass}
\alias{opiPresent_for_Compass}
\title{Implementation of opiPresent for the Compass machine.}
\arguments{
\item{\code{stim}}{A list containing:
\itemize{
\item \code{duration} Presentation time (ms).
\item \code{level} Stimuli luminance (cd/m^2).
\item \code{responseWindow} Response window (ms).
\item \code{x} x co-ordinates of stimulus (degrees).
\item \code{y} y co-ordinates of stimulus (degrees).
}}

\item{\code{...}}{Parameters for other opiPresent implementations that are ignored here.}
}
\value{
A list containing:
\itemize{
\item \code{res} JSON Object with all of the other fields described in @ReturnMsg except 'error'.
\item \code{res.eyet} Time of (eyex, eyey) pupil from stimulus onset (ms).
\item \code{res.eyex} x co-ordinates of pupil at times eyet (pixels).
\item \code{res.eyey} y co-ordinates of pupil at times eyet (pixels).
\item \code{res.num_track_events} Number of tracking events that occurred during presentation.
\item \code{res.msg.time} Response time from stimulus onset if button pressed (ms).
\item \code{res.eyed} Diameter of pupil at times eyet (mm).
\item \code{res.time_resp} Time since 'epoch' when stimulus response is
received, or response window expired (ms).
\item \code{res.msg} Error message or a structure with the following fields.
\item \code{res.error} '0' if success, something else if error.
\item \code{res.msg.seen} '1' if seen, '0' if not.
\item \code{res.num_motor_fails} Number of times motor could not follow fixation movement during presentation.
\item \code{res.time_rec} Time since 'epoch' when command was received at Compass or Maia (ms).
}
}
\description{
This is for internal use only. Use \code{\link[=opiPresent]{opiPresent()}} with
these Arguments and you will get the Value back.
}
\details{
\code{duration} can take on values in the range \code{[200.0, 200.0]}.

\code{level} can take on values in the range \code{[0.0, 3183.099]}.

\code{responseWindow} can take on values in the range \code{[0.0, 2680.0]}.

\code{x} can take on values in the range \code{[-30.0, 30.0]}.

\code{y} can take on values in the range \code{[-30.0, 30.0]}.
}
\examples{
chooseOpi("Compass")
result <- opiPresent(stim = list(duration = 200.0, level = 100.0, responseWindow = 1500.0, x = 0.0, y = 0.0))

}
\seealso{
\code{\link[=opiPresent]{opiPresent()}}
}
