% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/O900.r
\name{opiPresent_for_O900}
\alias{opiPresent_for_O900}
\title{Implementation of opiPresent for the O900 machine.}
\arguments{
\item{\code{stim}}{A list containing:
\itemize{
\item \code{size} Stimulus size (degrees). Can be Goldmann Size I to V (or VI if
device has a big wheel)
\item \code{color} Stimulus color (degrees).
\item \code{t} List of Stimulus presentation times (ms). For STATIC, list must
be of length 1. For KINETIC, it must the same length and 'x' and
'y' co-ordinates minus 1 (Optional)
\item \code{lum} List of stimuli luminances (cd/m^2).
\item \code{w} Response window for STATIC (ms). (Optional)
\item \code{x} List of x co-ordinates of stimuli (degrees).
\item \code{y} List of y co-ordinates of stimuli (degrees).
\item \code{type} Stimulus type: STATIC or KINETIC.
}}

\item{\code{...}}{Parameters for other opiPresent implementations that are ignored here.}
}
\value{
A list containing:
\itemize{
\item \code{res} List with all of the other fields described in @ReturnMsg except 'error'.
\item \code{res.msg.eyey} y co-ordinates of pupil at times eyet (degrees).
\item \code{res.msg.eyex} x co-ordinates of pupil at times eyet (degrees).
\item \code{res.msg.time} Response time from stimulus onset if button pressed (ms).
\item \code{res.msg} Error message or a structure with the following fields.
\item \code{res.error} '0' if success, something else if error.
\item \code{res.msg.seen} '1' if seen, '0' if not.
\item \code{res.msg.x} x co-ordinate when oberver responded (degrees) for KINETIC.
\item \code{res.msg.y} y co-ordinate when oberver responded (degrees) for KINETIC.
}
}
\description{
This is for internal use only. Use \code{\link[=opiPresent]{opiPresent()}} with
these Arguments and you will get the Value back.
}
\details{
\code{size} can take on values in the set \code{{"gi", "gii",
      "giii", "giv", "gv", "gvi"}}.

\code{color} can take on values in the set \code{{"white", "red", "blue"}}.

Elements in \code{t} can take on values in the range \code{[0.0, 1.0E10]}.

\code{lum} can take on values in the range \code{[0.0, 3183.099]}.

\code{w} can take on values in the range \code{[0.0, 1.0E10]}.

Elements in \code{x} can take on values in the range \code{[-90.0, 90.0]}.

Elements in \code{y} can take on values in the range \code{[-90.0, 90.0]}.

\code{type} can take on values in the set \code{{"static", "kinetic"}}.
}
\examples{
chooseOpi("O900")
result <- opiPresent(stim = list(size = "GV", color = "null", lum = 3183.099, x = list(0.0),
                  y = list(0.0), type = "null"))

}
\seealso{
\code{\link[=opiPresent]{opiPresent()}}
}
