% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimYes.r
\name{opiPresent_for_SimYes}
\alias{opiPresent_for_SimYes}
\title{Always respond 'yes' immediately to any parameter.
No checking is done on the validity of \code{stim}.}
\usage{
opiPresent_for_SimYes(stim, ...)
}
\arguments{
\item{stim}{Any stimulus object or list.}

\item{...}{Any other paramters you like, they are ignored.}
}
\value{
A list contianing:
\itemize{
\item err, and error code that is always NULL
\item seen, which is always TRUE
\item time, which is always 0
}
}
\description{
Always respond 'yes' immediately to any parameter.
No checking is done on the validity of \code{stim}.
}
