% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImoVifa.r
\name{opiQueryDevice_for_ImoVifa}
\alias{opiQueryDevice_for_ImoVifa}
\title{Implementation of opiQueryDevice for the ImoVifa machine.}
\arguments{
\item{\code{}}{A list containing:}
}
\value{
A list containing:
\itemize{
\item \code{error} \code{TRUE} if there was an error, \code{FALSE} if not.
\item \code{msg} If \code{error} is \code{TRUE}, then this is a string describing the error.
If \code{error} is \code{FALSE}, this is an empty list.
}
}
\description{
This is for internal use only. Use \code{\link[=opiQueryDevice]{opiQueryDevice()}} with
these Arguments and you will get the Value back.
}
\examples{
chooseOpi("ImoVifa")
result <- opiQueryDevice()

}
\seealso{
\code{\link[=opiQueryDevice]{opiQueryDevice()}}
}
