% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimHenson.r
\name{opiQueryDevice_for_SimHenson}
\alias{opiQueryDevice_for_SimHenson}
\title{Returns a simple list.}
\usage{
opiQueryDevice_for_SimHenson()
}
\value{
A list with one element \code{machine} that is \code{"SimHenson"}.
}
\description{
Returns a simple list.
}
