% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimNo.r
\name{opiQueryDevice_for_SimNo}
\alias{opiQueryDevice_for_SimNo}
\title{Returns a simple list.}
\usage{
opiQueryDevice_for_SimNo()
}
\value{
A list with one element \code{machine} that is \code{"SimNo"}.
}
\description{
Returns a simple list.
}
