% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compass.r
\name{opiSetup_for_Compass}
\alias{opiSetup_for_Compass}
\title{Implementation of opiSetup for the Compass machine.}
\arguments{
\item{\code{settings}}{A list containing:
\itemize{
\item \code{fixShape} Fixation target type for eye. (Optional)
\item \code{fixCx} x-coordinate of fixation target (degrees): Only valid values
are -20, -6, -3, 0, 3, 6, 20 for fixation type 'spot' and -3,
0, 3 for fixation type 'square'. (Optional)
\item \code{tracking} Whether to correct stimulus location based on eye position. (Optional)
}}
}
\value{
A list containing:
\itemize{
\item \code{res} List with all of the other fields described in @ReturnMsg except 'error'.
\item \code{res.msg} The error message or a structure with the result of QUERY OPI command.
\item \code{res.error} '0' if success, something else if error.
}
}
\description{
This is for internal use only. Use \code{\link[=opiSetup]{opiSetup()}} with
these Arguments and you will get the Value back.
}
\details{
\code{fixShape} can take on values in the set \code{{"spot", "square"}}.

\code{fixCx} can take on values in the range \code{[-20.0, 20.0]}.

\code{tracking} can take on values in the range \code{[0.0, 1.0]}.
}
\examples{
chooseOpi("Compass")
result <- opiSetup(settings = list())

}
\seealso{
\code{\link[=opiSetup]{opiSetup()}}
}
