% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimGaussian.r
\name{opiSetup_for_SimGaussian}
\alias{opiSetup_for_SimGaussian}
\title{Does nothing.}
\usage{
opiSetup_for_SimGaussian(state)
}
\arguments{
\item{state}{Any object you like, it is ignored.}
}
\description{
Does nothing.
}
