% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimHenson.r
\name{opiSetup_for_SimHenson}
\alias{opiSetup_for_SimHenson}
\title{Does nothing.}
\usage{
opiSetup_for_SimHenson(settings)
}
\arguments{
\item{settings}{Any object you like, it is ignored.}
}
\description{
Does nothing.
}
